requirements.txt
confuse~=1.5.0
azureml~=0.2.7
python-dotenv~=0.19.0
setuptools~=52.0.0
mlflow~=1.20.2
matplotlib~=3.2.1
scikit-learn~=0.23.2
numpy~=1.21.6
pandas~=1.3.5
azureml-core~=1.49.0
azureml-interpret
azureml-pipeline
azureml-defaults
azure-cli